notebooks/20240929_test_beam/pyproject.toml
[project]
name = "20240929-test-beam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-beam[aws,dataframe,docs,test]>=2.59.0",
    "python-dotenv>=1.0.1",
]
